+++ nukeserversocket/controllers/houdini/scripts/python/app/main.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 
 from nukeserversocket.main import NukeServerSocket
 from nukeserversocket.utils import stdoutIO
+from nukeserversocket.controllers import BaseController
 from nukeserversocket.received_data import ReceivedData
-from nukeserversocket.editor_controller import BaseController
 
 
 class HoudiniController(BaseController):
