+++ nukeserversocket/server.py
@@ -9,7 +9,7 @@ from .logger import get_logger
 from .received_data import ReceivedData
 
 if TYPE_CHECKING:
-    from .editor_controller import BaseController
+    from .controllers import BaseController
 
 LOGGER = get_logger()
